+++ dw.js/src/dw.dataset.js
@@ -5,7 +5,8 @@
 dw.dataset = function(columns, opts) {
 
     // make column names unique
-    var columnsByName = {};
+    var columnsByName = {},
+        origColumns = columns.slice(0);
     _.each(columns, function(col) {
         uniqueName(col);
         columnsByName[col.name()] = col;
@@ -58,6 +59,11 @@ dw.dataset = function(columns, opts) {
 
         hasColumn: function(x) {
             return (_.isString(x) ? columnsByName[x] : columns[x]) !== undefined;
+        },
+
+        indexOf: function(column_name) {
+            if (!dataset.hasColumn(column_name)) return -1;
+            return _.indexOf(columns, columnsByName[column_name]);
         },
 
         toCSV: function() {
@@ -115,6 +121,15 @@ dw.dataset = function(columns, opts) {
             uniqueName(column);
             columns.push(column);
             columnsByName[column.name()] = column;
+            return dataset;
+        },
+
+        reset: function() {
+            columns = origColumns.slice(0);
+            columnsByName = {};
+            _.each(columns, function(col) {
+                columnsByName[col.name()] = col;
+            });
             return dataset;
         }